requirements.txt
gradio>=4.0.0
torch
diffusers>=0.30.0
transformers
accelerate
opencv-python
pillow
numpy
segment-anything
xformers